+++ src/Http/Controllers/Dominion/RealmController.php
@@ -36,9 +36,8 @@ class RealmController extends AbstractDominionController
         $wonderHelper = app(WonderHelper::class);
 
         $dominion = $this->getSelectedDominion();
-        $round = $dominion->round;
 
-        if ($round->realmAssignmentDate()->addMinutes(5) > now()) {
+        if ($dominion->round->realmAssignmentDate()->addMinutes(5) > now()) {
             $request->session()->flash('alert-warning', 'You cannot access this page until realm assignment is finished.');
             return redirect()->back();
         }
@@ -49,7 +48,7 @@ class RealmController extends AbstractDominionController
 
         $isOwnRealm = ($realmNumber === (int)$dominion->realm->number);
 
-        if ($round->start_date > now() && !$isOwnRealm) {
+        if (!$dominion->round->hasStarted() && !$isOwnRealm) {
             $request->session()->flash('alert-warning', 'You cannot view other realms before the round begins.');
             return redirect()->route('dominion.realm', (int)$dominion->realm->number);
         }